+++ src/behaviors/EntryProductBehavior.php
@@ -5,24 +5,25 @@ namespace davidhirtz\yii2\cms\shopify\behaviors;
 use davidhirtz\yii2\cms\models\ModelCloneEvent;
 use davidhirtz\yii2\cms\models\Entry;
 use davidhirtz\yii2\cms\shopify\validators\ProductIdValidator;
-use davidhirtz\yii2\shopify\models\Product;
-use davidhirtz\yii2\shopify\models\queries\ProductQuery;
+use davidhirtz\yii2\shopify\models\traits\ProductRelationTrait;
 use Yii;
 use yii\base\Behavior;
 
 /**
- * EntryProductBehavior extends {@see Entry} by providing 'product_id` validation. This behavior is attached on
+ * EntryProductBehavior extends {@see Entry} by providing `product_id` validation. This behavior is attached on module
  * bootstrap by {@see Bootstrap}.
  *
  * @property Entry $owner
  */
 class EntryProductBehavior extends Behavior
 {
+    use ProductRelationTrait;
+
     public function events(): array
     {
         return [
-            Entry::EVENT_CREATE_VALIDATORS => 'onCreateValidators',
-            Entry::EVENT_BEFORE_CLONE => 'onBeforeClone'
+            Entry::EVENT_CREATE_VALIDATORS => $this->onCreateValidators(...),
+            Entry::EVENT_BEFORE_CLONE => $this->onBeforeClone(...),
         ];
     }
 
@@ -31,18 +32,9 @@ class EntryProductBehavior extends Behavior
         $this->owner->getValidators()->append(new ProductIdValidator());
     }
 
-    /**
-     * @param ModelCloneEvent $event
-     */
-    public function onBeforeClone($event): void
+    public function onBeforeClone(ModelCloneEvent $event): void
     {
         Yii::debug('Setting product_id to null on cloned entry.', __METHOD__);
         $event->clone->setAttribute('product_id', null);
-    }
-
-    public function getProduct(): ProductQuery
-    {
-        /** @noinspection PhpIncompatibleReturnTypeInspection */
-        return $this->owner->hasOne(Product::class, ['id' => 'product_id']);
     }
 }