+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0 (Nov 2, 2023)
+
+- Moved source code to `src` folder
+- Changed namespace of `davidhirtz\yii2\cms\shopify\widgets\grid\columns\ProductIdColumn` to `davidhirtz\yii2\cms\shopify\widgets\grids\columns\ProductIdColumn`
+- Upgraded `davidhirtz/yii2-shopify` to version `^2.0`
+
 # 1.0.8 (Nov 2, 2023)
 
 - Locked `davidhirtz/yii2-shopify` to version `^1.1`, upgrade to version 2.0 to use the latest version of this package